SONIC-1227 rename to BpkDividedCard
