+++ app/src/main/java/net/skyscanner/backpack/demo/data/ComponentRegistry.kt
@@ -34,6 +34,7 @@ import net.skyscanner.backpack.demo.compose.CardStory
 import net.skyscanner.backpack.demo.compose.CheckboxStory
 import net.skyscanner.backpack.demo.compose.CollapsibleNavigationBarStory
 import net.skyscanner.backpack.demo.compose.ColorsComposeStory
+import net.skyscanner.backpack.demo.compose.DividedCardStory
 import net.skyscanner.backpack.demo.compose.DividerStory
 import net.skyscanner.backpack.demo.compose.ElevationComposeStory
 import net.skyscanner.backpack.demo.compose.FieldSetStory
@@ -43,7 +44,6 @@ import net.skyscanner.backpack.demo.compose.HeadingStyleStory
 import net.skyscanner.backpack.demo.compose.HeroStyleStory
 import net.skyscanner.backpack.demo.compose.HorizontalNavComposeStory
 import net.skyscanner.backpack.demo.compose.IconsStoryCompose
-import net.skyscanner.backpack.demo.compose.InventoryDividedCardStory
 import net.skyscanner.backpack.demo.compose.NavigationBarStory
 import net.skyscanner.backpack.demo.compose.NudgerStory
 import net.skyscanner.backpack.demo.compose.PanelStory
@@ -304,6 +304,7 @@ object ComponentRegistry {
         TAB_TITLE_COMPOSE composeStory { ComposeDialogStory() },
       ),
     ),
+    "Divided Card" composeStory { DividedCardStory() },
     "Divider" composeStory { DividerStory() },
     "FieldSet" composeStory { FieldSetStory() },
     "Flare" story NodeData(
@@ -338,7 +339,6 @@ object ComponentRegistry {
         TAB_TITLE_COMPOSE composeStory { HorizontalNavComposeStory() },
       ),
     ),
-    "Inventory Divided Card" composeStory { InventoryDividedCardStory() },
     "Map Markers" story NodeData(
       { children -> SubStory of children },
       mapOf(
