+++ docs/compose/DividedCard/screenshots/default.png
